go*: run bash in loop by default
By default, when a container is run in a PWD environment, it is done so
without any args, and therefore assumes the default ENTRYPOINT and CMD.
Therefore, we set the defaults CMD to be one that runs bash in an
eternal loop so the user can't accidentally leave the in-browser
terminal.

Also shift any handling of .netrc in side the relevant guard.