all: multi-arch image support
All development of PWG had previously been done on amd64 machines. This
meant we had baked in lots of assumptions about base image digests.

For things to work fast and well on non-amd64 machines, we need to
switch to a platform-agnostic approach.

Switch the docker build process to target linux/arm64 and linux/amd64.

Also add a basic GitHub workflow that builds all the images.